Resolved conflicts (#91)
* Replaced logging with common trait

* Added explanatory comment

* Commit before merge

* Resolved some errors

* Resolved other errors and continued implementation of search rooms

* Implemented get user participations in room service

* Refactored

* Implemented get user participations in webapp service and web client

* Renamed method in web client

* Refactored responses adding method endWith to RichHttpServerRequest

* Optimized imports

* Added test

* Fixed error im query

* Added query param

* Deleted .gitkeep files

* removed test template from blank page

* Default routing to '/' when a room is deleted

* Added native window provider

* Fixed run.sh

* Added logback support (which permits to set the logging level programmatically)

* Fixed some errors in merge

* Resolved merge conflicts

+++ build.sh
@@ -2,14 +2,63 @@
 
 set -e
 
-cd dcs-client
-./build.sh
-cd ..
-
-echo "Removing build directories"
-for dir in `find . -maxdepth 2 -type d -name "build"`; do
-    # `rm -r` fails if the directory to delete is empty.
-    rm -r -- $dir || rmdir $dir
+while [[ $# -gt 0 ]]; do
+    key="$1"
+    case $key in
+        angular)
+            ANGULAR=1
+            shift # past argument
+            ;;
+        clean)
+            CLEAN=1
+             shift # past argument
+            ;;
+        check)
+            CHECK=1
+             shift # past argument
+            ;;
+        stage)
+            STAGE=1
+            shift # past argument
+            ;;
+        test)
+            TEST=1
+            shift # past argument
+            ;;
+        *)    # unknown option
+            echo "Unknown option: '$1'"
+            usage
+            exit 1
+            ;;
+    esac
 done
 
-./gradlew clean check test stage
+if [ ! -z ${ANGULAR+x} ]; then
+    cd dcs-client
+    ./build.sh
+    cd ..
+fi
+
+if [ ! -z ${CLEAN+x} ]; then
+    echo "Removing build directories"
+    for dir in `find . -maxdepth 2 -type d -name "build"`; do
+        # `rm -r` fails if the directory to delete is empty.
+        rm -r -- $dir || rmdir $dir
+    done
+    ./gradlew clean
+fi
+
+if [ ! -z ${CHECK+x} ]; then
+    echo "Checking sources"
+    ./gradlew check
+fi
+
+if [ ! -z ${TEST+x} ]; then
+    echo "Executing tests"
+    ./gradlew test
+fi
+
+if [ ! -z ${STAGE+x} ]; then
+    echo "Staging"
+    ./gradlew stage
+fi
